refactor(grfbrowser): Extract commands.go from main.go
Move screenshot and command handling to separate file:
- GUIState: State export structure for testing
- Command: Remote command structure
- captureScreenshot: Screenshot capture with OpenGL
- showNotification: Brief overlay messages
- dumpState: JSON state export
- checkAndExecuteCommand: Command file polling
- executeCommand: Command execution dispatch

Part of ADR-012 Stage 0: GRF Browser refactoring.
